move RequestResult to client.js
